+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - fix: unify storage path for `appGroupIdentifier` across targets ([#356](https://github.com/PostHog/posthog-ios/pull/356))
+- fix: do not call flags callback with invalid flags ([#355](https://github.com/PostHog/posthog-ios/pull/355))
 
 ## 3.26.2 - 2025-06-03